+++ bootmgr-rs-slint/src/editor.rs
@@ -1,0 +1,60 @@
+//! The optional basic editor for modifying [`Config`]s.
+
+use alloc::vec::Vec;
+use bootmgr_rs_core::config::{Config, editor::ConfigEditor};
+use slint::{Model, ModelRc, SharedString, ToSharedString};
+
+/// The basic editor
+#[derive(Default)]
+pub struct Editor {
+    /// If the editor is currently editing.
+    pub editing: bool,
+
+    /// The [`ConfigEditor`].
+    pub edit: ConfigEditor,
+}
+
+impl Editor {
+    /// Creates a new [`Editor`].
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Load an editor from a config.
+    pub fn load_config(&mut self, config: &Config) {
+        self.edit = ConfigEditor::new(config);
+        self.editing = true;
+    }
+
+    /// Save an editor to a config.
+    pub fn save_config(
+        &mut self,
+        config: &mut Config,
+        fields: &ModelRc<(SharedString, SharedString)>,
+    ) {
+        self.save_fields(fields);
+        self.edit.build(config);
+        self.editing = false;
+    }
+
+    /// Get the fields of the config.
+    pub fn get_fields(&self) -> ModelRc<(SharedString, SharedString)> {
+        let fields: Vec<_> = self
+            .edit
+            .fields()
+            .iter()
+            .map(|(x, y)| (x.to_shared_string(), y.to_shared_string()))
+            .collect();
+
+        ModelRc::from(&*fields)
+    }
+
+    /// Save the fields to the config.
+    pub fn save_fields(&mut self, fields: &ModelRc<(SharedString, SharedString)>) {
+        for (label, value) in fields.iter() {
+            if self.edit.go_to_field(&label) {
+                self.edit.update_selected(&value);
+            }
+        }
+    }
+}